Avoid silly class name
